Move classes to respective folders
